Move helper functions to textHelpers
