+++ src/components/RegionModal/Geography.js
@@ -3,8 +3,6 @@ import { Container, Header, Divider, List } from 'semantic-ui-react';
 import JSONTree from 'react-json-tree';
 import { theme } from '../styles/RegionModalStyles';
 import {
-  generateSubListItem,
-  generateValueItem,
   generateArea,
   generateElevation,
   generateEnvironment,
@@ -13,7 +11,11 @@ import {
   generateNaturalHazards,
   generateMaritimeClaims,
 } from './GeographyHelpers';
-import { generateTableList } from '../../helpers/textHelpers';
+import {
+  generateSubListItem,
+  generateValueItem,
+  generateTableList,
+} from '../../helpers/textHelpers';
 
 const Economy = ({ data }) => {
   const {
