+++ src/helpers/textHelpers.js
@@ -79,6 +79,54 @@ export const generateList = list => (
   </List>
 );
 
+export const generateTextItem = (obj = {}) => {
+  const entries = Object.entries(obj);
+  if (entries.length === 0) return '';
+  const [title, text] = entries[0];
+  if (typeof text !== 'string') return '';
+  return (
+    <List.Item>
+      {title && <strong>{capWithSpacing(title)}:</strong>}
+      {text && ` ${text}`}
+    </List.Item>
+  );
+};
+
+export const generateSubListItem = (obj = {}) => {
+  const entries = Object.entries(obj);
+  if (!entries) return '';
+  const [title, text] = entries[0];
+  if (typeof text !== 'string') return '';
+  return (
+    <List.Item>
+      {title && <List.Header>{capWithSpacing(title)}</List.Header>}
+      <List>
+        {text &&
+          text
+            .split(';')
+            .filter(x => x)
+            .map((item, idx) => <List.Item key={idx}>{item}</List.Item>)}
+      </List>
+    </List.Item>
+  );
+};
+
+export const generateValueItem = (obj = {}) => {
+  const entries = Object.entries(obj);
+  if (entries.length === 0) return '';
+  const [title, valObj] = entries[0];
+  if (typeof valObj !== 'object') return '';
+  return (
+    <List.Item>
+      {title && <strong>{capWithSpacing(title)}:</strong>}
+      {valObj.value && ` ${numScale(valObj.value)}`}
+      {valObj.units && ` ${valObj.units}`}
+      {valObj.date && ` (${valObj.date})`}
+      {valObj.note && ` (${valObj.note})`}
+    </List.Item>
+  );
+};
+
 export const generateTable = (table, title) => {
   const items = Object.keys(table[0]);
